Remove mknod from library.
This is all nowhandled in the SPL.  When HAVE_GPL_ONLY_SYMBOLS is defined
because the module license is set to GPL we can leverage the udev interfaces
which will create the device.  When HAVE_GPL_ONLY_SYMBOLS is not defined
we resort to an ugly hack and the SPL invokes a usermode helper to create
to missing device file for us in /dev/.  Using GPL version modules is off
course preperable when possible.