Libspl XDR implementation reorganization
Moved xdr_control() function from static inline in xdr.h in to a
new xdr.c file which was added to the libspl build.  Additionally,
I have removed the 'xdr_bytesrec' typedef which shares the same
name as the struct.  This is what Solaris does, but it's just asking
for trouble.  It has been replaced with an 'xdr_bytesrec_t' typedef.