Add NONE option for --with-linux and --with-spl.
This is used when you need to configure the project but you don't
actually intend to build it.  Thus you don't really need access to
either the kernel or spl headers and symbols.  At Livermore I use
this when I only intend to use the 'make dist' or 'make srpm' target.