Revert to previous behavior zdb fails when unable to open /dev/zfs.
For the moment I have added an error message to the failure path to
make it clear what happened.  I have also changed the zdb ASSERT to
a VERIFY so we always catch the failure.  For now we will just always
ensure the module stack is loaded, longer term we need something a
little more flexible.