Added zpios-sanity.sh script to validate common configurations.
This script was added to provide a simple way to check that zpool
layers correctly on all the standard linux block device types.
It's still a little fragile if there a hiccup in say the md or
lvm tool chain but aside from that it works well.

The 'make check' target now also calls this script in a safe mode
which only operates on files and loopback devices.  To check other
block devices types is must be explicitly run by hand because it
will overwrite various block devices.